# Conflicts:
#	src/main/java/com/example/accounts/Account.java
#	src/main/java/com/example/accounts/AccountController.java
#	src/main/java/com/example/accounts/Transfer.java

+++ src/main/java/com/example/accounts/AccountController.java
@@ -57,7 +57,7 @@ public class AccountController {
             @PathVariable String accountId,
             @RequestParam String transferId,
             @RequestParam String destAccId,
-            @RequestParam BigDecimal amount) {
+            @RequestParam double amount) {
 
         logger.info("Transfer for: {}", amount);
